% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-limits.R
\name{find_token}
\alias{find_token}
\title{Find a token from token cache that can perform an API call}
\usage{
find_token(query, requests)
}
\arguments{
\item{query}{The API endpoint you want to query. See
\url{https://developer.twitter.com/en/docs/basics/rate-limits}.}

\item{requests}{How many requests you need to make.}
}
\value{
A token that make the API request while respecting
Twitter API rate limits.
}
\description{
Mostly meant for internal \code{socialsampler} use. Loads all
tokens from the token database, or the token registered
with \code{rtweet} if not tokens have been registered with
\code{socialsampler}, creates the corresponding bearer tokens
(see \code{\link[rtweet:bearer_token]{rtweet::bearer_token()}}), and returns the token
that can serve meet the requests. If no tokens meet
these requirements, waits until the minimum amount
until the API rate limits reset for a token in the token
database.
}
