% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{import_tokens_from_rtweet}
\alias{import_tokens_from_rtweet}
\title{Import tokens registered with rtweet into the socialsampler token database}
\usage{
import_tokens_from_rtweet()
}
\description{
Import tokens registered with rtweet into the socialsampler token database
}
\seealso{
Other tokens: 
\code{\link{get_all_tokens}()},
\code{\link{inspect_token_db}()},
\code{\link{register_token}()},
\code{\link{unregister_all_tokens}()}
}
\concept{tokens}
