% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{inspect_token_db}
\alias{inspect_token_db}
\title{Inspect the socialsampler token database}
\usage{
inspect_token_db()
}
\description{
Prints information on the number of tokens registered with
\code{socialsampler}. If you haven't registered an tokens,
sampling function will use the default \code{rtweet} token,
and barring this, will give an error that tokens have
been registered with neither \code{rtweet} or \code{socialsampler}.
}
\seealso{
\code{\link[rtweet:get_token]{rtweet::get_token()}}, \code{\link[rtweet:create_token]{rtweet::create_token()}},
\code{\link[rtweet:bearer_token]{rtweet::bearer_token()}}

Other tokens: 
\code{\link{get_all_tokens}()},
\code{\link{import_tokens_from_rtweet}()},
\code{\link{register_token}()},
\code{\link{unregister_all_tokens}()}
}
\concept{tokens}
