% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{register_token}
\alias{register_token}
\title{Register a Twitter authentication token with socialsampler}
\usage{
register_token(api_key, api_secret, access_token, access_secret)
}
\arguments{
\item{api_key}{Twitter application API key (character).}

\item{api_secret}{Twitter application API secret (character).}

\item{access_token}{Twitter access token (character).}

\item{access_secret}{Twitter accces secret (character).}
}
\description{
If you try to register a token already in the cache,
you'll get a warning. That is, you are protected from
creating duplicate entries in the token cache.
}
\details{
For more details on how to get tokens please read
\code{vignette("auth", package = "rtweet")}.
}
\seealso{
\code{\link[rtweet:create_token]{rtweet::create_token()}}

Other tokens: 
\code{\link{get_all_tokens}()},
\code{\link{import_tokens_from_rtweet}()},
\code{\link{inspect_token_db}()},
\code{\link{unregister_all_tokens}()}
}
\concept{tokens}
