% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{safe_get_favorites}
\alias{safe_get_favorites}
\title{Safely get the tweets favorite by a collection of users}
\usage{
safe_get_favorites(users, attempts = 1)
}
\description{
Safely get the tweets favorite by a collection of users
}
