% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{safe_get_followers}
\alias{safe_get_followers}
\title{Safely get follower list of Twitter users}
\usage{
safe_get_followers(users, n = 5000, page = -1, verbose = FALSE, attempts = 5)
}
\arguments{
\item{n}{Number of followers to return. Defaults to 5000, which is
the max number of followers returned by a single API
request. Twitter allows up to 15 of these requests every 15
minutes, which means 75,000 is the max number of followers to
return without waiting for the rate limit to reset.  If this
number exceeds either 75,000 or the remaining number of possible
requests for a given token, then the returned object will only
return what it can (less than n) unless retryonratelimit is set
to true.}

\item{page}{Default \code{page = -1} specifies first page of JSON
results. Other pages specified via cursor values supplied by
Twitter API response object. If \code{parse = TRUE} then the
cursor value can be extracted from the return object by using the
\code{next_cursor} function.}

\item{verbose}{Logical indicating whether or not to print messages.
Only relevant if retryonratelimit = TRUE. Defaults to TRUE,
prints sleep times and followers gathered counts.}

\item{attempts}{How many times should we attempt to access the Twitter
API before giving up? Defaults to 5.}
}
\value{
An edgelist with columns:
\itemize{
\item \code{from}: A character vector of node ids. This will match the
input format of \code{users}. That is, if \code{users} consists of
screen names, \code{from} will consist of screen names. If
\code{users} consists of user IDs, \code{from} will as well.
\item \code{to}: A character vector of node ids.
}
}
\description{
\code{safe_get_followers()} is a drop-in replacement for
\code{\link[rtweet:get_followers]{rtweet::get_followers()}} that automatically respects
Twitter API rate limits and coordinates use of
multiple Twitter access tokens that have been
registered with \code{\link[=register_token]{register_token()}}. At the moment
you can only request the friends list of 15 or
fewer users in a single call.
}
\details{
If no tokens have been registered with \code{socialsampler},
\code{safe_get_followers()} will look for a token registered
with \code{rtweet} and use that token instead, again
automatically respecting rate limits. If no tokens have
been registered with either \code{socialsampler} or \code{rtweet},
you'll get an error.
}
