% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{safe_get_timelines}
\alias{safe_get_timelines}
\title{Safely sample Twitter user timelines}
\usage{
safe_get_timelines(user, n = 100, max_id = NULL, home = FALSE, attempts = 5)
}
\arguments{
\item{user}{Vector of user names, user IDs, or a mixture of both.}

\item{n}{Number of tweets to return per timeline. Defaults to 100.
Must be of length 1 or equal to length of user. This number should
not exceed 3200 as Twitter limits returns to the most recent 3,200
statuses posted or retweeted by each user.}

\item{max_id}{Character, returns results with an ID less than (that is,
older than) or equal to `max_id`.}

\item{home}{Logical, indicating whether to return a user-timeline
or home-timeline. By default, home is set to FALSE, which means
\code{get_timeline} returns tweets posted by the given user. To
return a user's home timeline feed, that is, the tweets posted by
accounts followed by a user, set home to TRUE.}

\item{attempts}{How many times should we attempt to access the Twitter
API before giving up? Defaults to 5.}
}
\value{
A tbl data frame of tweets data with users data attribute.
}
\description{
\code{safe_get_timelines()} is a drop-in replacement for
\code{\link[rtweet:get_timelines]{rtweet::get_timelines()}} that automatically respects
Twitter API rate limits and coordinates use of
multiple Twitter access tokens that have been
registered with \code{\link[=register_token]{register_token()}}. At the moment
you can only request the friends list of 15 or
fewer users in a single call.
}
\details{
If no tokens have been registered with \code{socialsampler},
\code{safe_get_timelines()} will look for a token registered
with \code{rtweet} and use that token instead, again
automatically respecting rate limits. If no tokens have
been registered with either \code{socialsampler} or \code{rtweet},
you'll get an error.
}
