% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-calls.R
\name{safe_lookup_users}
\alias{safe_lookup_users}
\title{Safely lookup Twitter users}
\usage{
safe_lookup_users(users, attempts = 5)
}
\arguments{
\item{users}{User id or screen name of target user.}

\item{attempts}{How many times should we attempt to access the Twitter
API before giving up? Defaults to 5.}
}
\value{
A tibble of users data.
}
\description{
\code{safe_lookup_users()} is a drop-in replacement for
\code{\link[rtweet:lookup_users]{rtweet::lookup_users()}} that automatically respects
Twitter API rate limits and coordinates use of
multiple Twitter access tokens that have been
registered with \code{\link[=register_token]{register_token()}}. At the moment
you can only request the friends list of 15 or
fewer users in a single call.
}
\details{
If no tokens have been registered with \code{socialsampler},
\code{safe_lookup_users()} will look for a token registered
with \code{rtweet} and use that token instead, again
automatically respecting rate limits. If no tokens have
been registered with either \code{socialsampler} or \code{rtweet},
you'll get an error.
}
