% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialsampler-package.R
\docType{package}
\name{socialsampler-package}
\alias{socialsampler}
\alias{socialsampler-package}
\title{socialsampler: Tools to facilitate sampling Twitter data}
\description{
Extends rtweet with additional functionality to gracefully
    handle multiple tokens, to automatically use bearer tokens
    for increased sampling capacity, and to automatically retry failed
    API requests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexpghayes/socialsampler}
  \item Report bugs at \url{https://github.com/alexpghayes/socialsampler/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

}
\keyword{internal}
