% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{register_token}
\alias{register_token}
\title{Register a Twitter authentication token with socialsampler}
\usage{
register_token(api_key, api_secret, access_token, access_secret)
}
\arguments{
\item{token}{A Twitter auth token token created by
[rtweet::create_token()].}
}
\description{
If you try to register a token already in the cache,
you'll get a warning. That is, you are protected from
creating duplicate entries in the token cache.
}
\details{
For more details on how to get tokens please read
`vignette("auth", package = "rtweet")`.
}
\concept{managing tokens}
