% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{unregister_all_tokens}
\alias{unregister_all_tokens}
\title{Remove all sampled users from the cache}
\usage{
unregister_all_tokens()
}
\description{
Will ask you for confirmation before proceeding because
of the how time consuming it is to sample users.
}
\concept{cache management}
